Format output keys to camelCase and input to snake_case
We want our API endpoints to respond with json data which
has camelCased key - a naming convention that seems more
natural to work with in JavaScript.

We also expect to receive data with camelCased keys, but want
to transform them to snake_case - a naming convention that seems
more natural to work with in Elixir.